--- README.md
@@ -1,16 +1,0 @@
-# ICT2206 MaLink Detector
- 2206 Project Github
- 
-# ICT2206-Web-Security-Assignment
-Repo for ICT2206 Assignment
-
-1. Design novel software that is broadly applicable to Web Security (Offensive/Defensive)
-2. Project proposal by 28 May 2359
-3. Project and relevant deliverables due 23 July 2359, consisting of:
-  - Source Codes/binaries (This repo)
-  - Detailed Report (follow IEEE Template, with references)
-  - Poster
-  - Video (5-10 min)
-  - Set of slides for 12 min group presentation
-
-Assigment weight: 35%
